chore: move groupings logic in products loader to a separate util
